+++ redis/test.js
@@ -339,6 +339,40 @@ tests.MULTI_EXCEPTION_1 = function() {
     });
 };
 
+tests.MULTI_8 = function () {
+    var name = "MULTI_8", multi1, multi2;
+
+    // Provoke an error at queue time
+    multi1 = client.multi();
+    multi1.mset("multifoo_8", "10", "multibar_8", "20", require_string("OK", name));
+    multi1.set("foo2", require_error(name));
+    multi1.set("foo3", require_error(name));
+    multi1.incr("multifoo_8", require_number(11, name));
+    multi1.incr("multibar_8", require_number(21, name));
+    multi1.exec(function () {
+        require_error(name);
+
+        // Redis 2.6.5+ will abort transactions with errors
+        // see: http://redis.io/topics/transactions
+        var multibar_expected = 22;
+        var multifoo_expected = 12;
+        if (server_version_at_least(client, [2, 6, 5])) {
+            multibar_expected = 1;
+            multifoo_expected = 1;
+        }
+
+        // Confirm that the previous command, while containing an error, still worked.
+        multi2 = client.multi();
+        multi2.incr("multibar_8", require_number(multibar_expected, name));
+        multi2.incr("multifoo_8", require_number(multifoo_expected, name));
+        multi2.exec(function (err, replies) {
+            assert.strictEqual(multibar_expected, replies[0]);
+            assert.strictEqual(multifoo_expected, replies[1]);
+            next(name);
+        });
+    });
+};
+
 tests.FWD_ERRORS_1 = function () {
     var name = "FWD_ERRORS_1";
 
@@ -363,6 +397,72 @@ tests.FWD_ERRORS_1 = function () {
     client.publish(name, "Some message");
     setTimeout(function () {
         client3.listeners("error").push(originalHandlers);
+        assert.equal(recordedError, toThrow, "Should have caught our forced exception");
+        next(name);
+    }, 150);
+};
+
+tests.FWD_ERRORS_2 = function () {
+    var name = "FWD_ERRORS_2";
+
+    var toThrow = new Error("Forced exception");
+    var recordedError = null;
+
+    var originalHandler = client.listeners("error").pop();
+    client.removeAllListeners("error");
+    client.once("error", function (err) {
+        recordedError = err;
+    });
+
+    client.get("no_such_key", function (err, reply) {
+        throw toThrow;
+    });
+
+    setTimeout(function () {
+        client.listeners("error").push(originalHandler);
+        assert.equal(recordedError, toThrow, "Should have caught our forced exception");
+        next(name);
+    }, 150);
+};
+
+tests.FWD_ERRORS_3 = function () {
+    var name = "FWD_ERRORS_3";
+
+    var recordedError = null;
+
+    var originalHandler = client.listeners("error").pop();
+    client.removeAllListeners("error");
+    client.once("error", function (err) {
+        recordedError = err;
+    });
+
+    client.send_command("no_such_command", []);
+
+    setTimeout(function () {
+        client.listeners("error").push(originalHandler);
+        assert.ok(recordedError instanceof Error);
+        next(name);
+    }, 150);
+};
+
+tests.FWD_ERRORS_4 = function () {
+    var name = "FWD_ERRORS_4";
+
+    var toThrow = new Error("Forced exception");
+    var recordedError = null;
+
+    var originalHandler = client.listeners("error").pop();
+    client.removeAllListeners("error");
+    client.once("error", function (err) {
+        recordedError = err;
+    });
+
+    client.send_command("no_such_command", [], function () {
+        throw toThrow;
+    });
+
+    setTimeout(function () {
+        client.listeners("error").push(originalHandler);
         assert.equal(recordedError, toThrow, "Should have caught our forced exception");
         next(name);
     }, 150);
@@ -811,7 +911,7 @@ tests.HLEN = function () {
             next(name);
         });
     });
-}
+};
 
 tests.HMSET_BUFFER_AND_ARRAY = function () {
     // Saving a buffer and an array to the same key should not error
@@ -828,7 +928,7 @@ tests.HMSET_BUFFER_AND_ARRAY = function () {
 // TODO - add test for HMSET with optional callbacks
 
 tests.HMGET = function () {
-    var key1 = "test hash 1", key2 = "test hash 2", name = "HMGET";
+    var key1 = "test hash 1", key2 = "test hash 2", key3 = 123456789, name = "HMGET";
 
     // redis-like hmset syntax
     client.HMSET(key1, "0123456789", "abcdefghij", "some manner of key", "a type of value", require_string("OK", name));
@@ -839,12 +939,23 @@ tests.HMGET = function () {
         "some manner of key": "a type of value"
     }, require_string("OK", name));
 
+    // test for numeric key
+    client.HMSET(key3, {
+        "0123456789": "abcdefghij",
+        "some manner of key": "a type of value"
+    }, require_string("OK", name));    
+
     client.HMGET(key1, "0123456789", "some manner of key", function (err, reply) {
         assert.strictEqual("abcdefghij", reply[0].toString(), name);
         assert.strictEqual("a type of value", reply[1].toString(), name);
     });
 
     client.HMGET(key2, "0123456789", "some manner of key", function (err, reply) {
+        assert.strictEqual("abcdefghij", reply[0].toString(), name);
+        assert.strictEqual("a type of value", reply[1].toString(), name);
+    });
+
+    client.HMGET(key3, "0123456789", "some manner of key", function (err, reply) {
         assert.strictEqual("abcdefghij", reply[0].toString(), name);
         assert.strictEqual("a type of value", reply[1].toString(), name);
     });